+++ admin-backend/src/main/java/com/example/backend/service/v2/SystemUserServiceV2.java
@@ -132,8 +132,9 @@ public class SystemUserServiceV2 {
             throw new BusinessException(BusinessErrorCode.USER_NOT_LOGIN);
         }
 
+        // TODO
         // 该用户角色没有权限删除用户
-        Integer loginUserRoleId = SessionUtils.getRoleId(session);
+        Long loginUserRoleId = SessionUtils.getRoleId(session);
         if (loginUserRoleId != 1 && loginUserRoleId != 2) {
             throw new BusinessException(BusinessErrorCode.OPERATION_NOT_ALLOWED);
         }
@@ -151,7 +152,7 @@ public class SystemUserServiceV2 {
         }
 
         // 要删除用户是管理员身份 不允许删除
-        Integer userToDeleteRoleId = userToDelete.getRoleId();
+        Long userToDeleteRoleId = userToDelete.getRoleId();
         if (userToDeleteRoleId == 1) {
             throw new BusinessException(BusinessErrorCode.OPERATION_NOT_ALLOWED, "不允许删除超级用户");
         } else if (!systemRoleService.canEmpowerTargetRole(loginUserRoleId, userToDeleteRoleId)) {
